requirements.txt
torch>=2.0.0
transformers>=4.35.0
huggingface_hub
datasets
anthropic>=0.39.0
openai>=1.0.0
google-generativeai>=0.8.0
together>=1.0.0
pyyaml
tqdm
pandas
matplotlib
seaborn
tenacity
python-dotenv
accelerate
sentencepiece
protobuf